Added Planets gravity and rocket spawning
There are some issues with gravity that don't look too nice, i.e. the
rocket sometimes gains a huge burst of speed when at the center of the
planet. I might fix this by making an indestructible center which will
fix a few issues. Also I have to make the input coords more consistent
with the co-ords libgx uses to draw the textures.